chatbot/requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
# LangChain 0.2.x stack (compatible with langchain-openai 0.1.12)
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.12
unstructured==0.14.4 # Document loading
onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.4.24
openai==1.30.5 # For embeddings
tiktoken==0.7.0  # For embeddings 
nltk==3.9.1 # For sentence tokenization in loaders/splitting